small pr to make firemode code easier | or, how firemode code was so fucking terrible that this pr touches like 20-30 files (#3061)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->




## About The Pull Request




Wanted to add pride flags for pride month, but had to get this 'small
pr' out the way. Oops.




tldr; as a player, this should not affect you. Like at all. Move along




This changes firemodes to not be fucking terrible codewise and to touch
as a developer. Yes, i HAD to touch gun.dm that much. That was how deep
rooted the rot was.


Also get's rid of the 0.4 second hardcoded fire delay, now pistols and
revolvers shoot as fast as they were intended. Great. This meant that i
had to touch a bunch of numbers i set to 0 out of frustration at it not
working, being completely unaware of the hardcoded fire delay.



This also fixes the E-40 to have laser full auto and to have an ammo
counter. I wasn't even trying to fix the E-40, but by unshittifying
firemode code i got it to work. Wow.



## Why It's Good For The Game


This should make the 3 upcoming gun expansions less ass to add.


<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Changing firemodes on guns now have new sprites
fix: The E-40 now works as intended!
add: The E-40 now has an ammo counter
balance: The P16 shoots slightly faster.
balance: .299 Eoehoma caseless has gotten a slight nerf, while the E-40
shoots slightly faster.
balance: Two E-40s max are obtainable from the black market should it
spawn. Blank market price cap of the E-40 has also increased

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: retlaw34 <58402542+retlaw34@users.noreply.github.com>