handrails (#3002)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This adds handrails, from Baystation - original PR that added them is
[here](https://github.com/Baystation12/Baystation12/pull/18474). This
doesn't actually add anything to any maps, but adds it as a structure
for mappers to use. I did a little bit of work by adding a crafting
recipe, and simplifying things by making it a chair subtype.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It's like. You can grab onto them. Makes Zero-G ships better, helps with
transits so people don't go awkwardly rushing to chairs or just flop to
the floor.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: fighterslam, comma
add: Handrails, rails on the wall that you can grab/buckle yourself to.
(*Not included in any maps with this PR, but craftable with 4 iron
rods.)
:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->