re:eggbox (#3074)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Resprites the eggbox, Done on behalf of @CoiledLamb 

## Why It's Good For The Game


![image](https://github.com/shiptest-ss13/Shiptest/assets/58402542/cb914d43-5b68-462d-b3ed-94bdc9009ff8)

## Changelog

:cl: CoiledLamb 
add: Resprites the eggbox
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: retlaw34 <58402542+retlaw34@users.noreply.github.com>