Removes Chaplain Magic (#2936)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Removes a lot of magic effects the chaplain like purifying cult runes
and healing people by smacking them on the head and immunity to cult
conversion
Makes the only chaplain amour witchhunter because templars are generic
(I know witchhunters are still pretty generic but it fits more with the
srm)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
We don't plan on ever using any of this content. 
Makes room for less cruft based chaplain content

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
del: Removed alot of chaplain stuff we will never use
del: The chaplain is now just as magic as the average person
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->