Concrete-ening: 3/4ths Edition (#3106)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Concrete and Hexacrete walls now follow the 3/4ths format. Tested and
confirmed it works as intended!

![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/62a585a0-81d5-438a-8a5c-667b92d4e967)
By @rye-rice 
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/23341cbc-547e-4db5-a440-e50c5a8bbecc)
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: Rye-Rice
imageadd: Rylie Was Forced To Resprite Concrete when she was 6!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->