Additional SUNS decals i missed, creates a decals folder (#3104)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Sorts out some unnecessary definitions of icon in turf decal subtypes,
creates a decals folder, adds four SUNS decals.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

I want to kill the leader of the Student-Union Association of
Naturalistic Sciences. Literally kill the leader of the Student-Union
Association of Naturalistic Sciences. No other person can come close to
relating how much i want to kill the leader of the Student-Union
Association of Naturalistic Sciences. There is no way you can convince
me to not kill the leader of the Student-Union Association of
Naturalistic Sciences. Killing the leader of the Student-Union
Association of Naturalistic Sciences could not possibly be anymore me. I
want to kill the leader of the Student-Union Association of Naturalistic
Sciences, and nobody can convince me otherwise.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
code: A lot of SUNS decals.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->