Small gun changes (#3048)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Makes the combat shotgun bulky instead of huge, and the compact combat
shotgun medium instead of bulky

Repaths the saber SMG to be a child of the skm carbine so the stock is
able to fold and changes

Adds automatic fire to the TEC-9 and makes it normal sized.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

The combat shotgun being huge is a holdover from station based server
balance. It being bulky now will make it be able to fit in armor vests.
Moves the compact combat shotgun down to normal as a result since it's
supposed to be the compact version.

The Saber and TEC-9 changes are so they're accurate to their
descriptions of having a folding stock and having a fast rate of fire
respectively. Ups the TEC-9 to normal sized as a result since having an
automatic pistol you can fit in your boot is probably not good.

## Changelog

:cl:

tweak: Saber SMG stock folds
tweak: TEC-9 Pistol is now automatic and normal sized
balance: Combat shotgun is now bulky, and the compact combat shotgun is
normal.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->