De-teshari-ing (Kepori Name Touch-up) (#3151)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Apparently, Kepori's namegen operates on the old namegen method that was
used back when Teshari were in the game. Its a good idea to, generally,
not let those linger for too long. So I went and made a name list for
some stored names and sets a basis for surnames (which is your parents'
given names connected together).

(also it fixes an old bit of lore thats been removed but who reads the
species blurbs on chargen)

## Why It's Good For The Game

Not having old names for a species we don't do anymore, cutting off the
"rollaways" bit, giving the basis for surnames for Kepori so they're not
single-namers anymore.

thgvr edit:
I fixed up the changelog for you

## Changelog

:cl:
add: Added Kepori names, surnames, and a list for the char gen to draw
from. Remember to update your character names!
del: Removed old Kepori name gen system and a bit of lore from the
species blurb
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->