The SUNS Panacea-class Infirm Vessel (#3112)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Replaces the Aegis-class with the Panacea-class Infirm Vessel. A
generalist with a central theme of roleplaying around inpatients.

![image](https://github.com/shiptest-ss13/Shiptest/assets/55075747/f26da6d5-baba-4921-9d66-6ac5929e8c51)

![image](https://github.com/shiptest-ss13/Shiptest/assets/55075747/e9b5c43f-0d17-442b-be1a-5af4e24cea3e)

![panaceaclass](https://github.com/shiptest-ss13/Shiptest/assets/55075747/257a7b65-c4f3-4d6d-b7f8-f750f6b16184)
(in game screenshot is 2 changes out of date.)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Replaces an old ship and uses our new assets.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added the Panacea-class
del: Removed the Aegis-class
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->