Reduces ship spawn limit for everything to 1 (#3166)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Reduces the spawn limit for every ship to 1, and makes the default limit
for new ships 1.

## Why It's Good For The Game

Nearly every ship was already set to a limit of 1. Making it the default
just saves time. Most of the handful of ships that still had a cap of 2-
mudskipper, li tieguai, and delta, for instance - were ships that would
just compete for pop or nudge each other out of their niche if
double-spawned, and a greater variety of ships are available to accept
players after existing ones fill up than there previously were.

## Changelog

:cl:
balance: reduced default ship spawn limit to 1.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->