Adds plastic flowers. (#3168)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a new item, a plastic flower. It can be reskinned into all of
flowers with an alt+click.
Adds that item into loadout, so chars can go and have flowers as part of
their drip.


![obraz](https://github.com/shiptest-ss13/Shiptest/assets/108196626/247a8051-3316-446d-9771-79c4d6bece6a)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More options for character creation.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added a plastic flower - a selection of flowers to wear. Added it
to the loadout too.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->