Removes cruft from the loadout. (#3213)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
List of removed items:
- Pet collar - meme item
- Tiki mask - very weird item
- Joymask - meme item
- Baloons - I don't think I've ever seen anyone with a baloon. Can
revert if it's somehow loved.
- Tinfoil hat - meme item
- Cat ears - meme item
- Horse mask - meme item
- Pirate hat - weird to have in the loadout
- galaxy suit - yeah, cuz average spessman can afford an animated suit
- tacticool turtleneck - use pants and a turtleneck undershirt


This also reflavors the Cybersun labcoat into a "translucent labcoat" so
it can stay in loadout until a unique coat is made for them.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Loadout has a bunch of items that it shouldn't have.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
del: Removed a bunch of meme, weird and unused items from the loadout.
tweak: Reflavored the Cybersun labcoat into a generic "translucent
labcoat"
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->