Increases the price of automatic weapons and some handguns (#3111)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Self explanatory. Also slightly nerfs the price on pistols to make them
more balanced compared to revolvers.

## Why It's Good For The Game

Automatics are EXTREMELY powerful and it's supposed to be a big deal to
equip everyone or most people in a crew with them. ESPECIALLY AR's.
Pistols one is more to just make revolvers more viable.

## Changelog

:cl:

balance: Doubled the price on every automatic gun and added 250 credits
to the price of every semi-auto pistol.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Martinpachu <86135885+Martinpachu@users.noreply.github.com>
Co-authored-by: Theos <theubernyan@gmail.com>