Changes the ingredients of two Gezenan drinks to better reflect their descriptions. (#3209)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR changes the foodtypes of the Gezenan Dark Mead and Keh'Lu'Tex
Liqueur to more accurately reflect their descriptions.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This change is good for the game as it will enable Sarathi to enjoy
drinks which are referenced to be popular Gezenan drinks. Both of the
changed drinks are described to be made without the use of grain or
vegetable products. This PR simply adjusts their foodtypes to reflect
this and permit Gezena's majority population to enjoy native beverages.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: fixed the foodtype of two Gezenan drinks to match their
descriptions
add: added a taste description to Keh'Lu'Tex Liqueur
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->