Adds a new ruin: the Abandoned Listening Post (#3103)
## About The Pull Request

Adds a brand new ruin to Lava Worlds, a remaster of the old listening
post ruin now abandoned mid-ICW and searched by the Ramzi Clique for
valuables.

Major Loot:
- Stechkin, loaded + 1 spare magazine
- Safe containing Loadsamoney spawner and Secret Documents, plus folder
to hold documents in
- Blackbox 
- Telecommunications assembly, capable of receiving only
- C-20r in end room with no magazine and one spare unloaded magazine
- Scalpal, Cautery and Latex Gloves in the small infirmary

Contains for threat:
- 3 Pistol Ramzi Commandos
- 1 SMG Ramzi Commando 
- 1 Shotgun Ramzi Commando in the final room

StrongDMM Areas Screenshot:

![image](https://github.com/user-attachments/assets/3735ed3d-fda3-4528-845f-8124f6957fca)
StrongDMM No Areas Screenshot:

![image](https://github.com/user-attachments/assets/14c12ef1-f323-4614-9a95-e865d6f272ac)
In-Game Screenshot:

![image](https://github.com/user-attachments/assets/5abb80d7-1363-4f1f-95b3-97a7713a761e)
## Why It's Good For The Game

With the recent Ruincut more ruins are necessary to round out the
planets in-game, and replacing a poorly designed ruin from the past with
a higher-quality remake is something I like doing. This also adds an
interesting environment of an abandoned holding from the ICW, of which
not many exist as ruins despite the ICW being such a large event.

## Changelog

:cl: Cloudbreak
add: A new ruin, the Abandoned Listening Post.
add: Areas associated with the listening post.
fix: Single line in maps catalogue.
/:cl:

---------

Signed-off-by: firebudgy <153147550+firebudgy@users.noreply.github.com>
Co-authored-by: meem <75212565+meemofcourse@users.noreply.github.com>