Remaps the Kilo (#3199)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This remaps the Kilo.
**TODO:**
- [x] ship desc
- [ ] wiki page
- [x] Maurice
- [ ] processing console directionals

Hull is now made from metal instead of titanium. The manufacturer was
swapped from NT to Miskilamo. Crew areas were expanded. KA and Egun were
swapped for two DBs - have fun with this one. Wires got a tiny bit
messier. Starting cash was nerfed from 5500 to 3500. ORM was replaced
with a smelter setup.


![scrnshot1](https://github.com/user-attachments/assets/91611c45-9b0c-4f0b-a929-f0fdc485570c)

![StrongDMM-2024-07-19 13 24
26](https://github.com/user-attachments/assets/2a65ac1d-2aa8-43d9-9bb8-b9019177ed42)

![StrongDMM-2024-07-14 01 21
13](https://github.com/user-attachments/assets/4babaf21-6143-48cb-8e47-95e76f50da2f)

**JOB SLOT CHANGES:**
- No doctor - There never was a medical. Why were they here?
- One more deckhand - Apogee requested the ship to stay at 5 crew
members.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Kilo's ancient. This modernizes it.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
tweak: Remapped the Kilo-class Mining Ship
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->