Adds an anesthetics crate to cargo (#3254)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds a cargo crate containing an Anesthetics tank and a medical
mask for it, for 500 credits.

## Why It's Good For The Game

It's rather difficult to get good quality anesthetics in-game without
having to resort to ghetto chemistry, salvaging for it, or by hoping to
get lucky with the black market. This PR should give an option for those
that want anesthetics available in-round, in a way that makes sense for
a medical professional to acquire.

## Changelog

:cl:
add: an anesthetics crate
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->