Makes 8x50 boxes have exactly 5 rounds more. (#3242)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Mystery PR.

## Why It's Good For The Game
Makes ammo for the illestren slightly cheaper! Enough for it to be as
good economically as pistols i think.
## Changelog

:cl:
balance: 8x50 boxes now come with 25 ammo as opposed to 20.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->