Social Anxiety quirk fixes and tweaks (#3222)
## About The Pull Request

Changes some things about the social anxiety quirk to make it more
bearable to deal with.

- Removes the ability for it to silence you, which would eat messages
players tried to send while roleplaying.
- Decreases the severity of stuttering across the board, and prevents it
from removing characters in the message.
- Psicodine now actually suppresses the main effects of social anxiety
while it is active, making the medicine actually usable once more.

## Why It's Good For The Game

Players were unable to use this quirk due to a variety of reasons. It
would randomly silence you around people, give you an exaggerated
stutter that made your runechat impossible to read, and couldn't even be
suppressed because it's treatment medication was bugged and had very
little use outside of it.

Now the quirk is a little more managable, and players should be able to
take it and roleplay with it again. I may mess with the stutter further
if needed as I still think its a little extreme in some cases.

## Changelog

:cl:
balance: the Quirk Social Anxiety no longer silences you randomly
fix: Stuttering is now far less extreme and will not delete characters
fix: Psicodine now actually suppresses the effects of social anxiety
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->