Adds player-accessible cryobeds to the indie outpost. (#3277)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Closes #2829
Replaces one of the four storages with cryogenics.

![scrnshot1](https://github.com/user-attachments/assets/b797b1bf-f694-43d9-bfb9-b09a6b9f34d7)

![StrongDMM-2024-08-13 22 26
31](https://github.com/user-attachments/assets/397acee4-bca2-482f-9f69-29a5b3a3e9de)

![StrongDMM-2024-08-04 15 39
42](https://github.com/user-attachments/assets/324a266a-b446-4b7b-8e68-e099ade8f94b)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Being able to cryo on the outpost is good in case Stuff happens.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Ihejirka space outpost now has player-accessible cryogenics.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->