Back In Black (Market) - Black Market Dead Drops and Stock Cycling (#3249)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

### The Catalogue
- The Black Market's catalogue now refreshes every hour with new
choices. The stock for individual entries will not refresh however if
bought. Check in every now and then!
-  Adds the tech and ammo tab
- Entries can be set to have weighting in order to be more or less
likely to appear. Items that recently appeared will be less likely to
show up again, and vice-versa for items that haven't. Particularly rare
or powerful items are exempted from weighting to stop them showing up
too often.
- Kills some items, adds some new items and adjusts some prices and
percentages.
- All blackmarket items are added to the market on generation, and
unavailable items are hidden.
- Paired items are now handled by a list so items can have multiple
paired items.

### Items
- Reflavours the ethereal disco grenade and makes it work on elzu. 
- Makes the powerfist actually work again. The gasused datum removed
itself before it could do the pressure checks.
- Adds a var (show_rigged) to power cells to whether they display
they're going to explode or not. This is a surprise tool that will hurt
you later.

### Delivery Changes
- Adds a new delivery method, Dead drops. The black market will drop off
your package somewhere near the ruin on the planet you're on. If there
isn't a ruin, or a viable spot to land, it'll send it somewhere random
and it's up to you to find it. A generally consistent way to get your
packages
- Makes launch more reliable, and your orders will no longer bounce off
your ship back into the void.
- The LTSRBT has been removed from the outpost market. The board lives
in the Black Market as an occasional spawn in the tech category, or can
be crafted with 2 BS crystals, a bank card, 5 duct tape, a circuit board
and a network card. The cost for bluespace shipping is now 100 credits
up from 50.

### Mechanical and Code Stuff
- Dynamic Encounters keep track of what map templates/ruins were spawned
in them in a list. Used for Dead dead drops so they can calculate a
landing spot, and should hopefully be useful in the future if we need to
get info about the ruins for whatever reason.
- adds get_block_portion and get_position_in_margin to mapzones.dm
- Makes the black market UI a bit wider so all the delivery options fit
on the menu nicely
- adds a parameter for get_overmap_object_by_locatiion to check for the
planet you're on exclusively rather than the ship


## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
The main intent of the PR was to mitigate the reliance of the LTRSBT as
a medium to effectively interact with the black market and was just a
1000 credit buy in fee effectively. It wasn't helped by the Launch was
unreliable to the point of being useless, and it feels really bad to
make a 10k purchase just to turn up empty handed cause it bounced off
your ship by a trajectory out of your control.

Having other viable methods of interacting with the black market should
hopefully mitigate the issue. Not to mention it felt a bit weird in
retrospect you could buy the Super Spooky Illegal Machine at the
outpost. It also doesn't pigeonhole you into asking your captain for an
LTSRBT if you want to smuggle in gear .

The Black Market stock cycling was something that seemed like people
wanted, so it here it is. Weighting also adds a bit more fine control on
what appears.

Rylie asked me to reflavour the disco grenade, and I thought it'd be
funny so here it is too.

## Changelog

:cl:
add: Black Market Stock Cycling and item weighting
add: New black market stock
add: Tech and Ammo tab
add: Dead Drops
add: LTRSBT moved to the Black Market Catalogue
add: LTSRBT Crafting Recipe, 2 BS crystals, a bank card, 5 duct tape, a
circuit board and a network card
add: Reflavours the disco grenade
add: variable for powercells to show if they're rigged
del: Some old blackmarket stock
balance: LTRSBT shipping cost up to 100 credits
balance: Launch delivery more reliable
fix: Powerfist works again
spellcheck: fixed a few typos in the black market
code: Dynamic Overmap encounters store what ruins were spawned
code: get_block_portion and get_position_in_margin methods in
mapzones.dm
code: Pair items now handled by a list
refactor: Black Market item stocking
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Gristlebee <56049844+Gristlebee@users.noreply.github.com>
Co-authored-by: Theos <theubernyan@gmail.com>