Renames the Mauler exosuit + exosuit guns (#3244)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Renames the Mauler and three exosuit guns to avoid confusion and make
their weapon class and expected damage more obvious.

- Mauler is renamed to Touro (honestly temporary, given the eventual
exosuit rework)
- FNX-99 incendiary carbine has been renamed from Hades to Phoenix
- Ultra AC 2 has been renamed to UMG-2 Machine Gun
- LBX AC 10 has been renamed to LBX-10 "Scattershot" Heavy Shotgun

## Why It's Good For The Game

people keep buying the Hades exosuit carbine expecting a Hades laser
carbine. also we're getting a Mauler machine pistol and talking about
that and the exosuit is going to get confusing.

## Changelog

:cl:
refactor: Renamed the Mauler exosuit and several exosuit weapons to
avoid confusion
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->