Eoehoma Caseless is now available on the black market (Plus the E-40 is very slightly cheaper) (#3318)
## About The Pull Request

Adds Eoehoma Caseless boxes to the game, available only in the black
market. Upper limit of the E-40 is now 10000 instead of 15000 (its still
on average 100% more expensive than a standard assault rifle)

## Why It's Good For The Game

It's a cool weapon but you never get to use the ballistic part more than
a few times. And I'm sure theres plenty of ammunition leftover for use
to sell on the black market.

## Changelog

:cl:
add: .299 caseless is now available in the black market for the E-40
balance: E-40 is now slightly less expensive
fix: .299? .229? No more typos
/:cl: