Adds two new lever action rifles. (#3243)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds the Absolution Lever Action Rifle, chambered in 357, and
Conflagration Lever Action Shotgun. Pictured below.

![imagen](https://github.com/user-attachments/assets/5294ed7e-5d64-44dd-90b9-10d14a19e58b)

![imagen](https://github.com/user-attachments/assets/dcffaaca-ce82-48eb-b2fa-423204542f22)

![imagen](https://github.com/user-attachments/assets/62b9fdd1-f0f7-4a7b-977a-1da4362690ba)

Spriting by RyeRice, Names by Valorium, Descriptions by Apogee, and
coding by me!!

## Why It's Good For The Game
Upgrades to the .38 lever action rifle that provide a nice variety, with
the lever-action shotgun being an inbetween of a double-barrel and pump
action shotgun and the 357 being more expensive and with a smaller ammo
capacity but with a lot more power.
## Changelog

:cl:
add: To-do once i get the proper names and all
/:cl:

---------

Signed-off-by: Martinpachu <86135885+Martinpachu@users.noreply.github.com>
Co-authored-by: Theos <theubernyan@gmail.com>