Fixes the "pulse carbine" on whitesands_surface_camp_saloon again (#3331)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Fixes the missing sprite on a varedited gun on the map. Also changes the
amount of money on the map since it was brought to my attention that it
had 14k credits lying about. Adds a few gunner hermits for good measure.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Invisible guns are bad. oh and ruins generally shouldnt have over 10k
credits unless they warrant it.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: The "pulse carbine" in the sandplanet saloon has a sprite again
tweak: Added some extra mobs to the sandplanet saloon to balance out the
money in it
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->