New items and rebalances for Outpost Cargo Exploration Tab + Underground Scanner detection change (#3328)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Rebalances the prices of old packs in the exploration tab, and adds some
new items.
- Lava Exploration Kit: Price: 1500 -> 500, adds two pocket
extinguishers
- Ice Exploration Kit: Price: 1500 -> 500, swaps out the boots for ice
hiking boots, adds two winter coats
- Jungle Exploration Ki:  Price: 750-> 500, adds two lanterns
- Bluespace shelter capsules: Price: 3000 -> 500, only contains a single
capsule now

Adds some new packs:
- Lavaproof rods crate, 200 credits
- Binoculars 200 credits
- Deep core drill scanner, 250 credits
- Anomaly neutralizer, 250 credits
- GPS, 100 credits

You can now click a vein with a deep core scanner on harm intent so the
scanner no longer picks it up, or hit it again to readd it to tracking.

The deep core scanner fits in explorer webbing storage and explorer
suits.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

A lot of the exploration packs were comically overcosted for what they
were. For 1500, the ice exploration pack got you.... two pickaxes, boots
that didnt protect from the ice, and cosmetic goggles. For 1500 you
could get two actual mining crates which would serve you a lot better.
The new prices should hopefully make them more worthwhile purchases.

The new packs are meant to be generally useful items to aid in
exploration.
Lavaproof rods are a useful, if niche nice to have if you're penned in
by lava.
Binocs and GPS are good for general exploration.
Anomaly neutralizers are handy for difficult anomalies that hard or
impossible to approach normally. Case in point:


https://github.com/user-attachments/assets/cabbdb99-d425-4227-863b-ffdd6549fff1

Adds the deep core mining scanner as an individual purchase so you don't
have to pay a 1000 credit buy in for the drill if you only intend to do
missions.

Detectability changes for veins are meant to be QOL, and streamline
finding the type of vein you want, as you can now remove unwanted veins
from being tracked instead of wandering aimlessly hoping to get close
enough to another vein for the scanner to lock on.

## Changelog


:cl:
add: Adds binocs, lavaproof rods, gps, deep core mining scanner, anomaly
neutralizer to exploration cargo
add: You can hit veins with a deep core scanner on harm intent to stop
it tracking, and hit it again to readd it.
add: Deep core scanner fits in explorer webbing and explorer suits.
balance: Outpost Cargo exploration tab
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->