Adds Inteq wintercoats and hoodies to the uniform lockers on Inteq ships (#3344)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Adds the inteq wintercoat and hoodie to the Talos, Colossus, Vaquero and
Valor uniform lockers.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

If you're cold, the Sarathi are cold too.

The Inteq wintercoat is only obtainable as an alt pref-option to the
armor vest for enforcers, and the hoodie isn't mapped anywhere as far as
I'm aware. It'd be nice to able to see them in game more often for
aesthetic options and it's a nice flavorful thing to have to keep warm
with.

## Changelog

:cl:
add: Inteq wintercoats and hoodies in the Talos, Colossus, Vaquero and
Valor uniform lockers
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->