Makes Ship Turrets Connectable Without Areas (#3386)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Hacks some stuff onto the code for turrets and turretid machines so that
they can see turrets outside the area (thanks fallcon). Applies this to
every single ship so that ship turrets are a little less weird.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Extendable + Useful for mappers. Also less area gore.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
code: Ship Turrets can now be attached to control consoles outside the
area they are in. please look at ship maps for examples on how.
add: Every single ship with turrets now does the above
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->