mudskipper and shetland window fix (#3426)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Added windows to the mudskipper and shetland's engines as well for
consistency see #3434 for the Kilo fixes too.

Also seperated the Shetland's engine rooms so both buttons don't open
both sides.
Now the button in each engine room will open only it's respective side.

## Changelog

:cl:
fix: Added windows to the mudskipper and shetland's engines.
fix: Adjusted the blast doors which open on the Shetland's engines.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->