Sawn variants for the illestren and improvised shotgun now properly have their stats set (#3427)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

If the illestren's sawn off accuracy is unreasonably high I'd like to
know because it's TERRIBLE


## Changelog

:cl:
fix: sawn off illestren/improvised shotgun stats are now consistent if
they are spawned in
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Theos <theubernyan@gmail.com>