Cabinet Refactor (#3378)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Mildly refactors fireaxe cabinets to allow them to be subtyped with
different items. Also adds an is_wielded proc to most weapons that can
be wielded to prevent potential weirdness with this down the line.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Cabinets are soulful
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: Erika Fox
code: Fireaxe cabinets have been repathed, and now function as a more
general cabinet object. please report any inconsistencies with behavior
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>
Co-authored-by: FalloutFalcon <falloutfalconplays@gmail.com>
Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>