skhrinks gun cases (#3481)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
i put it in the washing machine for too long 


![image](https://github.com/user-attachments/assets/e91b4291-0ce3-4149-aa73-1d3f76a95529)

## Changelog

:cl:
imageadd: Shrunk down the sprite of guncases
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->