Removes A Lot Of Cruft (#3447)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Removes the following:
Swarmers
Bloodcult
Soulstones
Everything Devils
Meteor Gamemode
Area/shuttle

- [x] Compiles and runs on my local

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I'm fairly sure none of this stuff fits regardless of if it works with
the game. And area/shuttle is annoying to see in sdmm.

God I hate it when you start pulling one thing out and it's easier to
just pull the rest out than to adjust it to work in the event that
someone might use it for an event once.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
code: A large amount of cruft has been deleted.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->