Circuit Board Resprite (#3497)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR redoes the visuals for the main machine/computer boards that can
be printed off of any circuit imprinter.


![image](https://github.com/user-attachments/assets/3e557ce0-bef5-4e35-be3a-4a07a0c79f9f)

More will come in the future.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Our previous ones were quite a mess in comparison, and it's a small
little thing to keep us occupied while I work on other, larger sweeping
changes. I've been meaning to PR it for a while in any case.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy
imageadd: We now have improved visuals for circuit boards!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->