Resprites The "Braid" Haircuts (#3510)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Before I continue to explain the PR, I kindly ask for this to not be
ported to any other codebases. This is merely a request, and is
ultimately unenforceable as long as proper credit is given to the
authors. (Me in this case)

A continuation of me and @rye-rice's efforts with respriting every hair
on Shiptest. This one tackles the Braid hairstyles, both long and short!


![image](https://github.com/user-attachments/assets/a6a18236-1a1d-4b57-a7e7-c6ac46e5f246)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Another two down babey!!!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy
imageadd: Long (And Short) Braid hairstyles have been redone!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->