To Whom It May Concern - Bullet Writing (#3552)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

You can now write on bullets with a pen to give your ~~victim~~
recipient a lovely message when they're hit.


![image](https://github.com/user-attachments/assets/859b62ac-5597-40f2-bc70-59bb6091015e)


## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

Adds a little personalization to your death dealing, and I think it'd be
funny.
## Changelog

:cl:
add: Bullets can have a message written on them.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->