Makes the Cobra not just a better Mongrel (and fixes a typo) (#3437)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Reduces the size of the C-20r and Cobra-20's magazines to 24 rounds,
identical to the Mongrel. This also means you get two clean reloads out
of a box of .45 ammunition instead of having one mag and _almost_ a
second mag. Also, fixed the "civillian" typo in the cargo listings for
scarborough guns.

## Why It's Good For The Game

not getting two reloads out of a box is just plain irritating and also
the cobra shouldn't just be a mongrel but better

## Changelog

:cl:
balance: Reduced Cobra-20 magazine size to 24 rounds
fix: removed a typo from Scarborough cargo catalog entries
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>