Raleigh-Class Mercenary Corvette (#3458)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
![StrongDMM-2024-10-06 19 02
02](https://github.com/user-attachments/assets/f44a57a7-ef44-4696-8c08-c00a2ba63367)
![StrongDMM-2024-10-06 19 02
03](https://github.com/user-attachments/assets/4b862d9e-13b7-486f-83ca-c5de470dc7ca)


![dreamseeker_qhkyYgbpdx](https://github.com/user-attachments/assets/2f4dd01a-527e-4e9f-bd70-ec9ad8653d73)

![dreamseeker_8jf3u0ZlwA](https://github.com/user-attachments/assets/d7e3b7ac-1e96-4b32-b3bc-a6e213c648db)

![dreamseeker_jJx71tKdzb](https://github.com/user-attachments/assets/16b5a426-d622-4059-926c-151ce0a41e61)

Adds the Raleigh Class Corvette - a century old design manufactured by
Kasagi-Fischer for yuppie solarians, and now retooled for usage as a
mercenary vessel.

Originally sold as a 'Safari Ship', or a 'Yacht' in some markets, the
strange shape and relative quality of the design made it very appealing
for usage by unaligned mercenaries forces, with the vessel capable of
carrying a small complement of supposedly professional soldiers.

The Raleigh supports a crew of:
1 - Captain
1 - Lieutenant
1 - Engineer
3 - Troopers
1 - Recruit (random guy we found)

The Raleigh's gear includes:
1 - E40
1 - Cobra-20
2 - Ringnecks
Assorted Machetes
The Yuri Room
The Gambling Room
The Emergency Medical Hologram


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Adds a combat/rp focused ship for mercenaries that's |not| inteq. Think
of all the possibilities.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Kasagi-Fischer has begun a new production run of the venerable
Raleigh-class Corvette. See your purchaser's manual for more
information!
/:cl:
<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Erika Fox <94164348+Bjarl@users.noreply.github.com>