Adds Nanotrasen, Vigilitas and N+S stamps! (#3641)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![obraz](https://github.com/user-attachments/assets/fd5c49bf-72e1-49d4-827f-e7182af60792)

Maps all that crap on both NT ships and NT ruins.
**Generic stamps are for non-command jobs.**
NT logo is temporary until the rename happens, so it sucks.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Faction stuff or whatever.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added Nanotrasen, Vigilitas and N+S stamps.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->