Removes ORM from non-mining/salvage ships. (#3235)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Removes the ORM from Beluga, Box, Schmiedeberg, Colossus, Skipper and
Kansatsu.

Aspawn ships, Hyena and Gecko still have it, because:
- Hyena is a mining/salvage ship, so it needs it.
- Gecko is a salvage ship, so it needs it.
- I've been told not to touch aspawn ships.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
They have no reason to have an ORM! They are not mining/salvage-centered
ships, so they can totally go to the outpost to refine their ore!
(it's a Dwayne buff in disguise :3)
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
del: Beluga-class, Box-class, Schmiedeberg-class, Colossus-class,
Skipper-class and Kansatsu-class lost their Ore Redemption Machines.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->