click cooldown hud (#3461)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

A port of this pr
https://github.com/DaedalusDock/daedalusdock/pull/732

Adds a bar to the bottom of the screen that shows click cooldown

https://github.com/user-attachments/assets/8824ba1a-2596-4dcf-ae16-2dafdd380666


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I can tell when i can attack next
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: FalloutFalcon, Kapu1178
add: New bar above your items to show your attack cooldown
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: Sun-Soaked <45698967+Sun-Soaked@users.noreply.github.com>