Carpentry Tweaks -- Allows easels and cabinets to be crafted, and fixes various bugs with painting frames (#3632)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds the ability to craft easels and cabinets, and it fixes the
spawned structure type when painting frames are used from being the
generic type to the library type. This PR also fixes a bug when placing
painting frames, as shown below.

| before | after |
|--------|-------|

|![dreamseeker_tpNvZO4I62](https://github.com/user-attachments/assets/95a26ded-9c47-4e75-9b79-9b5f5f82250f)|![dreamseeker_Yep9pLj1oV](https://github.com/user-attachments/assets/39b9cc72-a4d9-4517-a14f-a92204faf117)|

## Why It's Good For The Game

Cabinets and easels _should_ be craftable, so as to keep parity with
other materials. It also provides an avenue to pursue more low-tech
solutions when one is going for a more rustic aesthetic.

Additionally, crafted painting frames have had a minor problem with not
displaying saved paintings when one is placed in them because the type
of frame that typically gets mapped is not the same type that can be
crafted. This should fix that, _and_ a bug associated with pixel shifts
for the frame structure when it's initialized.

## Changelog

:cl:
add: Added ability to make cabinets and easels
fix: fixed painting frame persistence and initialization bugs
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->