NGR Kali Andhi-class Destroyer (#3554)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![kali_andhi_ingame](https://github.com/user-attachments/assets/dff8226f-0cfd-4086-be04-34941df001ab)

![kali_andhi](https://github.com/user-attachments/assets/71eaa7d3-ac03-4721-b0c4-001e0d686acc)

![kali_andhi_areas](https://github.com/user-attachments/assets/16497551-4328-49ee-932d-247a80079168)


Adds the Kali Andhi-class, a new class of NGR destroyer that makes up
the bulk of the NGR's young fleet. This class's activity in securing
NGR's territorial claims has made them an unwelcome sight for the PGF
especially. The Kali Andhi is a combat ship, comparable to the Crying
Sun or the Colossus, featuring good armament and plenty of space for the
crew to interact. Crew count is as follows:

- 1 NGR Captain
- 1 Lieutenant; Commands the Operatives, organizes and executes away
missions
- 1 Ensign; Mans the bridge, manages communications, coordinates the
support crew
- 3 Operatives; Soldiers, plain and simple. Can be armed with a
Sidewinder, a Hydra DMR, or an Asp.
- 1 Wrecker; Organize and handle the storage bay, retrieve resources,
and occasionally scout for the Operatives
- 1 Mechanic; Operate and maintain the power and life support systems
- 1 Medical Doctor; Maintain the crew's general health, organize and
manage medical supplies, bag and tag corpses
- 2 Initiates; General-purpose muscle for cleaning and maintaining the
ship.

Also adds a set of NGR Tan floor decals, NGR wall and desk flags, and
tweaks a few of the NGR outfits.

![kali_salute](https://github.com/user-attachments/assets/98f17072-4b00-4c44-ab79-f74f744aa09c)

## Why It's Good For The Game

NGR needs a ship and the Hyena doesn't count.

## Changelog

:cl:
add: the NGR Kali Andhi-class destroyer
add: NGR floor decals and flags
balance: red space suit armor made equivalent to sec hardsuits
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Sun-Soaked <45698967+Sun-Soaked@users.noreply.github.com>
Co-authored-by: meemofcourse <75212565+meemofcourse@users.noreply.github.com>
Co-authored-by: Sun-Soaked <45698967+Sun-Soaked@users.noreply.github.com>