Twohanded weapons now support having 0 force (#3698)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

setting force wielded or unwielded to 0 would cause it to permanently
keep the other value since it skipped setting that value when
wielded/unwielded

## Why It's Good For The Game

why??

## Changelog

:cl:
code: weapons which can be wielded may now have a force of 0 when
wielded or unwielded
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Theos <theubernyan@gmail.com>