Replaces the Hyena with the Derecho (#3693)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![derecho](https://github.com/user-attachments/assets/950af567-b610-46f3-9de9-e84bed9fd7c4)

![derecho_areas](https://github.com/user-attachments/assets/bf379173-0409-4666-9d00-6a89065617c7)

![derecho_ingame](https://github.com/user-attachments/assets/ce331440-b280-4f59-b304-fbecf56c39f0)

Replaces the Hyena with the Derecho-class Auxiliary Wrecker, a mid-sized
New Gorlex salvage ship. Comes with more advanced engineering gear than
usual and a moderate array of weaponry including two Asp pistols and one
Bulldog shotgun. Features a crew of six players:
- Captain: You know what this is.
- Foreman: Direct the efforts of the Wreckers and the Mechanic.
- Wrecker x2: Explore, mine, and most especially salvage wrecks for
materials and valuables. Comes with a limited number of advanced salvage
supplies including power tools and a plasma cutter. One wrecker can wear
a toolbelt, the other gets a mining webbing.
- Mechanic: Maintain, repair, and improve the ship.
- Initiate x2: The starter role. Keep the ship clean, your crewmates
fed, and in general try to be useful while learning the ropes of the
faction and/or the game.

## Why It's Good For The Game

NGR needs more ships and the Hyena's been slated for replacement for a
while now due to its outdated and profoundly crusty design.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Replaced NGR Hyena with the NGR Derecho-class salvage ship
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->