Sound Enabling: Vox Screaming (#3778)
## About The Pull Request

Within the files there existed a sound effect for Vox screaming, but for
one reason or another it was never actually enabled for them. This fixes
that.

## Why It's Good For The Game

More species being able to scream in pain after getting an arm torn off
is good.

## Changelog

:cl:
add: Enabled thee screaming audio emote for Vox.
/:cl: