Adds NVG's to the outpost market (#3616)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds Night vision goggles to the outpost, for 1000 credits a pop.

## Why It's Good For The Game

Currently as it is, night vision goggles are inaccessible player wise
unless your ship spawned with them, or you had found RnD in one of the
few ruins left behind with them. With this PR, it adds them to the
outpost cargo, which allows players to buy them without such needs.

Night vision goggles add to PvP a different field of combat with other
hostile vessels, which include ambushes or as we have seen with a recent
round, potential traps for the opposing side to fall into. Another
aspect is their use in exploration and potential PvE encounters, like
deep drills in caves or asteroids, or ruin exploration for ruins that
either start with low/no power or have no functional lights for the
explorers to see.

## Changelog

:cl:
add: Night vision goggles are now purchasable via the outpost for 1000
credits
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->