Auxmos update and potential fix (incredibly mergeable) (#3764)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Updates to the latest version of auxmos, and switches (temporarily) to a
branch with a fix I added to auxmos to potentially stop the runtimes
we're getting. Here's hoping at least

## Why It's Good For The Game
These runtimes are the bane of everyone

## Changelog

:cl:
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->