Adds names & descriptions for Frontiersman and Ramzi mobs (#3686)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds names and descriptions to the Frontiersman and Ramzi mobs, making
it a little easier to recognize which mob is doing what in chatlogs and
giving them a little more (possibly too much) flavor if you shift click
them. I might have cooked too hard on the miniboss/elite mobs.

## Why It's Good For The Game

ramzis being named Ramzi Clique Operative regardless of whether they're
punching you with the force of a sponge or turning you to mist with a
Bulldog is a little annoying, and their descs are extremely minimal

## Changelog

:cl:
add: Ramzi and Frontiersman mobs now have more names and descriptions
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->