Fixes the switchblade letter opener and machete becoming invisible (#3767)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

switchblade now uses on/off sprite & has no world icon because it doesnt
have a world icon
letter opener now doesn't use world icons because it's stored in a
completely seperate file and doesn't have one either

machete now doesn't try to change its sprite on wielding because it
doesnt need to do that and also doesn't have the required sprite

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: some weapons are now more visible than they weren't supposed to be
(which is none)
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->