Stale-checks flaky test issues (#3811)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Basically just makes flaky test fail issues stale after a week, then
closes them another week afterwards. My reasoning is that we're going to
end up with a bunch of flaky fails that were one-time (from issues
introduced in a PR and then fixed, for instance) or were fixed without
closing the issue. I feel like this is long enough, and if it does
happen again, it SHOULD re-open the same issue.