Removes the S.W.A.T. Suit from cargo. (#3805)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removes the cargo listing for the S.W.A.T. Suit, The item has not
removed due to concerns regarding subtyping.
## Why It's Good For The Game

This might be a little controversial, but it's what has to be done, and
the marine armor is already here to replace it. I hate the swat suit so
much, i have been getting constant requests every single fucking day to
remove it, so here it is. Good thing too, this thing is so repulsive i
can't look at it for two seconds without throwing up over my monitor.
One time i was playing a medship and one of my patients had the fucking
gall to be wearing one. I shot his ass thrice and got banned for two
months. Worth it.

## Changelog

:cl:

del: The cargo listing for the SWAT suit.
/:cl: