Moves the Brazillab ruin to Sandplanet and fixes the Trabuco's inhand sprites (#3375)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Remaps the Brazillab ruin to fit on whitesands, because on iceplanet all
of the hermits in it would die to the cold and the ruin looked ugly
there anyways.
Also fixes the trabuco's inhands that were apparently broken.
Ruin has been given a serious touchup, but otherwise remains similar to
its iceplanet counterpart.


![image](https://github.com/user-attachments/assets/16517740-302b-499e-b3ee-241bc3685d24)

<details>
<summary>SDMM Pictures</summary>

![StrongDMM-2024-11-16 00 30
10](https://github.com/user-attachments/assets/c0c7ae32-ce94-4f62-b343-e1ba29578d56)

![StrongDMM-2024-11-16 00 30
13](https://github.com/user-attachments/assets/35d0af51-926b-4623-8272-5d7537b9bd76)


</details>

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This ruin always was a better fit for sandplanet, and now has had a
serious touch-up so it better fits with our increasing quality
standards.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Remapped and moved the Brazillab ruin to Sandplanet.
fix: Fixed the Trabuco's inhand sprites.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: zevo <95449138+Zevotech@users.noreply.github.com>