Blocks items meant for other slots for being beak-holdable for Kepori (#3816)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

I'm sorry to do this and make it even more niche, but it IS kind of
jank. This also solves the issue of things being equipped to the wrong
slot when you hit the quick-equip or quick-swap key.

Fixes: #3657

## Why It's Good For The Game

Prevents any other oversights like welding goggles from being doable by
Kepori. Hopefully this doesn't block TOO many items, I even added a few
other slots that should be pretty harmless to equip items flagged as
such, but if there's any exceptions that should be made, I'm open for
feedback.

## Changelog

:cl:
tweak: Kepori can't hold items in their beak that are (allegedly) meant
to be equipped to other slots.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->