Factional Cargo, done in a day (#3758)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Split off from Overmap Expansion # 5

This was done in a day.

Yes. This Thing that has been anticipated for 2-3 years was done in a
day.

How did I even accomplish this?

While attempting to unhardcode cargo/clean up outpost code, I noticed
that factional cargo was actually really easy to do. So I did it. If I
really wanted to be fancy, I could take 12 months to code cargo from
scratch. But this was easier. So I did it.

huge thanks to @firebudgy for coding support and adding more stuff to
factional cargo instead of this just being a bare framework

## Why It's Good For The Game

![November 12 2024
462](https://github.com/user-attachments/assets/49e8e086-767b-4f8b-abbf-375e4d61d822)
![November 12 2024
461](https://github.com/user-attachments/assets/6777d7d5-a7d1-48d7-9eb1-57685123dae1)
![November 12 2024
460](https://github.com/user-attachments/assets/95df2b56-f499-4895-8600-28dbfebae145)

## Changelog

:cl: rye-rice, firebudgy 
add: Factional cargo.
del: the remminants of the P16 as it was already completely removed in
Serene Sporting
balance: Sporter is now 400 credits.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: rye-rice <58402542+rye-rice@users.noreply.github.com>
Signed-off-by: Mark Suckerberg <mark@suckerberg.gay>
Co-authored-by: retlaw34 <58402542+retlaw34@users.noreply.github.com>
Co-authored-by: Mark Suckerberg <mark@suckerberg.gay>