Reverts Hydrogen pipes in hangars, adds hydrogen pumps to nt ice and indie outpost (#3867)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Reverts hydrogen pipes and exchange terminal in hangars
Adds hydrogen pump and exchange to main outpost level

![image](https://github.com/user-attachments/assets/a6467124-19c4-4f80-9db9-212b9a6ae77d)

![indiechanges](https://github.com/user-attachments/assets/d498c22e-6bd2-4ab8-ab1c-4c4c02e056b2)

known bug, if a second outpost is spawned with admin tools, the volume
pump breaks and needs to be re-wrenched. no idea how to fix it. This
used to occur on all hangars so this should be an improvement?

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Atmospherics bugs make pumps not work in the most recent hangar, among
other jank. Also, the fuel pipelines weren't great either.
Hydrogen needs more work still, but at least it'll be obtainable without
having to work around obscure bugs.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: added hydrogen pump and exchange to ntice and indie outposts
del: reverts hangars, removing hydrogen pump, pipes, and exchange from
indie and ntice hangars
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->