Energy Shotgun (E-SG 500) De-shittifying (#3858)
## About The Pull Request

The energy shotgun (or the IoT shotgun, E-SG 500, etc.) has been a
genuinely, truly awful firearm within the game for as long as I can
remember. It shot lasers that tickled people, had almost no range
compared to ballistics, had four shots before running out and was an
energy weapon which meant reloading was a hassle or even impossible. You
could barely kill goliaths with it!

This changes it to have marginally increased range, to be slightly more
damaging and marginally more accurate. Ideally, this makes people
actually consider using the damn thing rather than it being used less
than the E-11 in combat because it's more valuable for the cells inside
the gun than the gun itself.

It now has 10 shots but given it's an energy weapon it can afford having
2 more shotgun shots due to the hassle of reloading.

## Why It's Good For The Game

I tried using this shitty thing in combat once and it travelled three
goddamned tiles before fizzling out. The few times I did hit anyone with
this piece of garbage it did less than an E-11 and with the spread it
meant I was more likely to shoot my ship's pet through the glass than I
was to actually hit any reasonable amount of shots on someone even point
blank. I brought this thing out on a lava world and the legions didn't
even attack me because even if I stood next to them I'd run out of ammo
before killing them and they just stood there and laughed at me with
their stupid fucking skulls just like Etherbor executives laughed me out
of their board room for proposing sawing off their BG-16s in order to
fit them in a briefcase and so after that i went home with my E-SG 500
and had a home invader and tried to use the damn thing as a club but it
snapped in half and by the time i managed to fix the thing he was
running off with my xbox and when i tried to shoot him the spread of the
laser shells broke my window and my TV and barely even fizzled his back
i tried to pawn the gun off but the pawn shop owner told me not even a
landfill would accept this piece of shit so i turned to my pastor to
exorcise my spirits from the possession of madness the e-sg 500 brought
upon my soul and during it the gun tried to shoot me but even controlled
by the spawn of satan it was so shit it only burned a hole in my pocket
before its battery died and the pastor just laughed at me for

## Changelog

:cl:
balance: Adjusts the E-SG to hopefully be used more often.
/:cl:

---------

Signed-off-by: firebudgy <153147550+firebudgy@users.noreply.github.com>