SKM-24u LMG Modification on the Black Market (#3871)
## About The Pull Request

Adds the SKM-24u LMG Modification to the black market, for on average
about 1000 above standard SKM sale price. Given the statistical
similarities to the regular SKM minus it's having of a bipod and the
non-guarantee of drum magazines, this felt right.

Also fixes bipods to actually work at all after the code for them was
just outright removed in a PR I can't care to find.

## Why It's Good For The Game

New rare weapons variety to the black market, and a defensive option in
the form of a bipod-attached LMG should make for some interesting
situations.

## Changelog

:cl:
add: An intercepted shipment of CM-40 barrels has led to an influx of
SKM-24u LMGs to the black market. Report sightings to your nearest CLIP
representative.
fix: Bipods actually work now.
/:cl:

---------

Signed-off-by: firebudgy <153147550+firebudgy@users.noreply.github.com>