Tegu-class Communications Freighter (#3016)
## About The Pull Request
This PR adds a brand new ship to Nanotrasen's post-rework roster.

<details><summary>Images</summary>


![68747470733a2f2f616666656374656461726330372e626c6f622e636f72652e77696e646f77732e6e65742f6d6462322f696d616765732f3336373235323037352f33333231373933373937312f612f5f6d6170735f73687574746c65735f6e616e6f74726173656e5f](https://github.com/user-attachments/assets/a2c95f8d-fab7-4f16-ba47-ce1b549fd6b3)

**images below are outdated**

![image](https://github.com/shiptest-ss13/Shiptest/assets/118859017/6a41eddd-fcbe-4630-b1b9-814d3f966671)


![image](https://github.com/shiptest-ss13/Shiptest/assets/118859017/39594bbe-7837-4d49-a21f-442e42af5e2d)

</details>

The Tegu-class is an engineering freighter equipped with a full
telecommunications setup and a subshuttle. It is built to not only
facilitate communications between Nanotrasen-aligned vessels in the
sector, but also foster several smaller networks for other vessels with
a small array of relays. The ship also comes equipped with a small cargo
department, an engineering department, a small communications room, and
a supply office for the QM.

This PR also adds a new design disk for printing telecommunications
parts.

To-do:
- [x] Ship description.
- [ ] In-game testing.
- [x] Add security role.
- [ ] Mining facilities.

## Why It's Good For The Game
more ship

## Changelog

🆑goober3, Cloudbreak
add: The year is 2070. The Tegu-class Communications Freighter has just
been fullmerged.
add: A new subshuttle, the Skink-class Cargo Runner.
/🆑