Riot shield variants now reduce recoil when equipped (#3862)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Default riot shields will provide a -2 to unwielded recoil for any gun
fired while the shield is in your offhand. This primarily buffs SMGs and
pistols since anything with significant unwielded recoil have a
considerable amount and anything with significant unwielded spread will
still have that unwielded spread

## Why It's Good For The Game

Improves the riot shield by allowing people to use it while also using a
gun that isn't a shotgun or laser

## Changelog

:cl:
balance: Holding a riot shield will now reduce the recoil of guns
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->