Allows the ion thruster to be detached from exosuits + Adds it to the outpost market (#3892)
## About The Pull Request

As is the title. Not only allows exosuit ion thrusters to be detachable
from exosuits if attached (Since there was one ruin that had such not
attached to a suit), but also adds it to the Outpost market for almost
150% increase in price to the old RCS thruster.

## Why It's Good For The Game

The RCS thruster honestly sucked ass to be real with you. It not only
siphoned air exceptionally fast, you got only a few tiles in space
before being left in a vacuum. This'll add the objectively superior
version to the outpost, albeit for a relatively steep price to account
with how much better it is compared to the gas powered one in terms of
use.

## Changelog

:cl:
add: Adds the Ion thruster pack for Exosuits to the outpost market, for
2000 credits
balance: Ion thrusters can be detached from exosuits now
/:cl: