Maptainer Mapcut 2024 (#3840)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This pull request removes the following ships from the code:
* independent_beluga
* The concept is cool, and the effort placed into the ship is
commendable, but it does not work in practice
* independent_box
* This ship is old, with a lack of factional play and any other gimmick
than healing. A new indie medical ship is in the works already
* independent_lagoon
* This is an old, very large BarRP ship. It is already adminspawn, might
as well finish the job
* independent_schmiedeberg
  * The Schmie is old and specializes in a very shallow mechanic
* independent_tranquility
* This ship is explicitly excluded from the main gameplay loop and is
reliant exclusively on your ability to roleplay with your own crewmates
instead of the outside world. In other words, BarRP. This is also
possibly the oldest ship in this list
* inteq_hound
  * The Hound is old, aspawn, nobody uses it, and is already replaced
* nanotrasen_gecko
* The Gecko has some awful pipe layouts, with purposeless maints. It is
also incredibly old
* nanotrasen_heron (and by extension nanotrasen_falcon)
* This ship is not used, works on old Nanotrasen lore standards, and it
is also way too fucking big with a meh layout
* nanotrasen_mimir
  * This map's concept is badly executed, and it is not used
* pirate_noderider (Jupiter)
* SEC is in lore hell, the concept is good but can be executed better.
Nobody uses this
* solgov_chronicle
* This ship suffers from a particularly bad case of Bridge Isolation,
and was a testbed for the other SolGov ships. Could do with a
replacement


This pull request removes the following ruins from the code:
* beach_fishing_hut
  * Might as well just start over 
* beach_float_resort
  * This is a gigantic map with nothing in it
* beach_ocean_town
  * This is a gigantic map with nothing in it
* icemoon_hydroponics_lab
  * jesus christ
* jungle_abandoned_library
* The ruin's concept is alright, but it is executed poorly and does not
entice its loot properly (people avoid it because it sucks to fight in
for not too worth it)
* rockplanet_nomadcrash
* rockplanet_harmfactory
* whitesands_surface_medipen_plant
  * This ruin is not fun
* astraeus
  * Holy shit this is old
* corporate_mining
* wasteplanet_lab
  * if you tell me this is a good ruin i will kill you
* wasteplanet_pandora

This also renames the ruins.dm file into _ruins.dm and moves it into the
datum/ruins folder.
___
<table><thead>
  <tr>
    <th>Faction<br></th>
    <th>Ships<br></th>
  </tr></thead>
<tbody>
  <tr>
    <td>Indie</td>
    <td>10</td>
  </tr>
  <tr>
    <td>Syndicate<br></td>
    <td>7</td>
  </tr>
  <tr>
    <td>Nanotrasen</td>
    <td>4</td>
  </tr>
  <tr>
    <td>Inteq<br></td>
    <td>4</td>
  </tr>
  <tr>
    <td>PGF</td>
    <td>3</td>
  </tr>
  <tr>
    <td>Pirates<br></td>
    <td>2</td>
  </tr>
  <tr>
    <td>SolCon</td>
    <td>2</td>
  </tr>
  <tr>
    <td>CLIP<br></td>
    <td>1</td>
  </tr>
  <tr>
    <td>Roumain</td>
    <td>1</td>
  </tr>
</tbody>
</table>

<table><thead>
  <tr>
    <th>Planet<br></th>
    <th>Ruins<br></th>
  </tr></thead>
<tbody>
  <tr>
    <td>Jungle<br></td>
    <td>6</td>
  </tr>
  <tr>
    <td>Lava<br></td>
    <td>6</td>
  </tr>
  <tr>
    <td>Space</td>
    <td>5</td>
  </tr>
  <tr>
    <td>Waste</td>
    <td>5<br></td>
  </tr>
  <tr>
    <td>Rock<br></td>
    <td>5</td>
  </tr>
  <tr>
    <td>Sand<br></td>
    <td>5<br></td>
  </tr>
  <tr>
    <td>Beach</td>
    <td>4</td>
  </tr>
  <tr>
    <td>Ice<br></td>
    <td>4</td>
  </tr>
</tbody>
</table>

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Thank you, Mapper, for so kindly and skillfully granting us the fruits
of your labor. But now I'm afraid you've outlived your usefulness. You
know just enough to be a threat to me, and, with that, the future of the
Shiptest Technical Test. I will not say Fasidomi misol, Mapper... as I
can assure you, we will Never. Meet. Again.

For real, though. A lot of these ships had it coming.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
del: Beluga-class, Box-class, Lagoon-class, Schmiedeberg-class,
Tranquility-class
del: Hound-class
del: Gecko-class, Heron-class (also Falcon-class), Mimir-class
del: Jupiter-class
del: Chronicle-class
del: Many ruins, mostly for being old, have been removed
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->