surface_camp_combination fixes and adjustments (#3809)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Some changes to spawners caused a spawner on this map to spawn a shitton
of exotic guns out in the open. This PR fixes the spawner on the map,
and adjusts the distribution of loot on the map to better reflect
current mapping standards.

Botany hut no longer has a free biogenerator and plant DNA manipulator
board. Also has less seeds.
Armory hut no longer has a Candor and surplus .45 box just sitting
inside of it. Now just the SKM-24v and polymer rifle.
Bar's unguarded double barrel shotgun has been replaced with an
improvised shotgun. Weird bag of coins was removed.
Removed the thermal eyepatch hidden in a cave. Replaced the armory
exotic gun spawner with a medical supplies spawner. Removed the
syndicate tool spawner sitting out in the open.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This was one of my earliest "good" ruins. Its loot distribution suffered
from me being a new to map balance at the time (me being a stupid
powergamer). It had a lot of powerful items hidden in weird spots. Now
the loot is less deranged and generally in line with other ruins of its
size.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: fixed an incorrect spawner on whitesands_surface_camp_combination
tweak: adjusted loot distribution on whitesands_surface_camp_combination
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: Gristlebee <56049844+Gristlebee@users.noreply.github.com>