Fixes the mech saw and improves it + Adds it to the outpost catalouge (#3869)
## About The Pull Request

As is title. This PR adds the mech saw to the outpost market and fixes
it at the same time while buffing it. Gristlebee was a big help in
fixing this, as mech code has me quite stumped and would probably give
me an aneurysm if I looked at it too long.

## Why It's Good For The Game

Mech saw was freezing if you sawed indestructible walls + the original
PR never added damage to it which caused it to get stuck on regular
walls too. Also adds the exosuit based salvage saw to the outpost market
as it was originally unobtainable without it being spawned or somehow
starting with it.

## Changelog

:cl:
add: Adds the exosuit based saw to the Outpost catalogue for 2000
credits
balance: Buffed the speed and deconstruction damage of the exosuit saw
fix: Exosuit saw no longer gets stuck on indestructible walls
/:cl: