Equip Delay (#3854)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
It now takes a moment to remove or equip clothing, triggering a do_after
that places the item in the appropriate slot once completed.

Base times are as follows:

![image](https://github.com/user-attachments/assets/962a8ebd-df09-4fb8-88de-2d65b443c055)

You can move while equipping some lighter items, but it slows you down.

Please thank @Kapu1178 for writing the better code in this pr, their
work can also be found in [This
PR](https://github.com/DaedalusDock/daedalusdock/pull/1093)

Comes with functionality for more sound effects being plugged in for
different types of items.

KNOWN ISSUES:

- [x] Equip noises can be spammed by repeatedly clicking during the
do_after

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
My immersion
More importance in picking the right gear for different situations. You
can no longer hotswap in and out of stuff like spacesuits instantly,
think a little more carefully about when to put them on.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: Kapu1178
add: There is now a customizable delay for equipping and unequipping
items from clothing slots.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->