Atoll-Class Update (#3956)
## About The Pull Request

Makes some various changes to the Atoll that I feel would be healthier
for the ship.

Namely:
- Replaces the surgical duffel bag with a surgery case.
- Adds health analyzers to each of the medical role's lockers, so they
won't have to take from the medkits.
- Buffs the contents of the food fridge to better compete with rations,
and also to be more carnivore friendly.
- Places a new subshuttle docking port outside the lobby's blast doors.
- Vaporizes a decal under a wall that went under my nose.

## Why It's Good For The Game

This ship's been tested and found wanting.

## Changelog

:cl:
fix: Makes some changes to the Atoll-Class's map file. It should feel
slightly better to play on.
/:cl: