Adds a CE access requirement to the wrecker's equipment room on the Derecho (#3998)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR tweaks the wrecker's equipment storage room's airlock so that it
requires either CE access or mining access in order to open.

## Why It's Good For The Game

This would allow for the Foreman aboard the Derecho to access the
wrecker's equipment, which I feel makes sense. @Apogee-dev , the
developer of the Derecho, is also fine with this change.

## Changelog

:cl:
code: modifies the access requirements to the wrecker's equipment room
aboard the Derecho so that the Foreman can access it
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->