Removes Bullet Absorbtion From Mobs (#3985)
this feature is dumb nothing is 'immune' to bullets

:cl:
balance: goldgrubs, fugu, and basilisks are now marginally less immune
to bullets.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>