Renames and rewrites the Odysseus and Gygax (#3946)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Changes the Odysseus and Gygax to be manufactured by Cybersun. Finally,
the company that specializes in robotics and cybernetics actually has a
presence in the exosuit market.
- Gygax is now the 500 Series, represented by the 501p (Gygax) and 515
EX (Dark Gygax).
- Odysseus is now the 200 Series, represented by the 202r, and
potentially more in the future.
- 515 EX has better power efficiency when running in overcharge mode.
The armor is still the same as the 501p, but you can run around like 30%
more without zapping your batteries to dust.

Originally, this was going to be packaged with a resprite/redesign of
the exosuits in question and some tweaks to their performance, but that
will have to wait until later because spriting exosuits is hell on
earth.

....I'll change the wide variety of paths associated with them Later(tm)

## Why It's Good For The Game

cybersun actually gets a presence in the exosuit market and we can
finally clean up some of the old outdated scraps of lore that were still
hanging around in relation to these two exosuits

## Changelog

:cl:
code: Renamed the Gygax and Odysseus and updated their lore
add: Dark Gygax (now 515 EX) has better efficiency when overloading its
legs.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->