Adds a Shaft Miner to the Harrier, removes the Janitor (Role) (#3972)
## About The Pull Request

Adds a Shaft Miner, and minorly adjusts the N+S cargo office to
accommodate. In return to keep the crew size identical, the Janitor's
slot has been removed and the room lowered to general access. Assistants
can take up the role if they would like to be a janitor.

![image](https://github.com/user-attachments/assets/6b771897-8d66-4a91-8fc1-8bc0e9f058de)

![image](https://github.com/user-attachments/assets/1059ce93-3d1c-4522-afcd-8c7c100cf970)


## Why It's Good For The Game

Inherently, the Harrier has an issue with integration into the gameplay
loop. A lack of combat equipment means that ruin raiding is not much of
an option, and most missions are either ruin raiding... or mining. But
the Harrier is unable to properly perform mining as almost all of it's
staff are civilian in nature. What's left for it is it's medical bay,
which is worse than dedicated medships and most would rather interact
with an Atoll than a Harrier.

Now, you may say, what about the Cargo Techs?
Cargo Technicians are wage-slaves to N+S, but their role is *not*
excavation. Nor is the Supply Director. These are warehouse workers.
There is a specialized role for that. Having spoken with some of the
most active Harrier captains, it's pretty clear that most of the time
the Cargo Technician is sent out anyways for mining since someone is
_needed_ to keep the ship running and supplied.

## Changelog

:cl:
add: N+S has deployed a new batch of Shaft Miners to Harrier-class
vessels.
del: Janitors have been reassigned from the Harrier for retraining.
/:cl: