Flare and floodlight tweaks (#3997)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR triples the lifetime of flares, and it increases their light
range from 7 to 12. It also modifies the light range list for
floodlights from (0, 5, 10, 15) to (0, 10, 15, 20).


![image](https://github.com/user-attachments/assets/a01f9581-d700-48d7-ae24-709d81dfb720)
This shows how much area the flares light up with this change.


https://github.com/user-attachments/assets/0208c6ff-b6d8-4ef8-95f0-4568bc875f4f
This video shows how much area the floodlights will swath in light with
this change.

## Why It's Good For The Game

I believe that this change will help flares be used more often, as this
will make them light up a greater area than they do right now, and it
will also increase the length of time for which they will burn. This PR
will also increase the effectiveness of floodlights.

## Changelog

:cl:
balance: increased flare lifetime and light range, and increased maximum
floodlight light range
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->