Ghost HUD Respawn Button (#4037)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Replaces the Jump To Mob HUD button with a respawn button.


![image](https://github.com/user-attachments/assets/d5adc2be-893f-4a06-97e3-4a513c597a7e)
Thanks @meikiato from the discord for the above text. I am not wordy.

![image](https://github.com/user-attachments/assets/23c53a46-4f17-40ad-b383-377ad41c8319)
The respawn button confirms your choice if you have a body to return to,
otherwise this confirmation will not appear.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Respawning is an important part of shiptest, whether you're checking
ships before the round or respawning after an unfortunate death. This
makes it a more accessible and obvious option for new players, and
requires less mucking about in the verb menu.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Replaced Jump To Mob on ghost HUD with a respawn option.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->