Adds an electric welding tool to Chief Engineer toolbelts and the ERT utility belt (#3999)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds an electric welding tool to chief engineering belts and ERT
utility belts.

## Why It's Good For The Game

The chief engineer's belt is currently missing a welding tool, and I
felt that it may make sense to use an electric welder here for that.

## Changelog

:cl:
add: Added electrical welding tools to the populate lists for CE belts
and ERT utility belts
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->