copys some tg code to fix the cleanbot fails (#4052)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
fixes the linters!
copy pastes the code from the tg cleanbot file at this point in history
https://github.com/Ryll-Ryll/tgstation/blob/b9318c10ae1ee4d1d88871c7c31c7f8265f8ee6b/code/modules/mob/living/simple_animal/bot/cleanbot.dm
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
No more failing test!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
refactor: lightly refactors the get path section of the cleanbot ai. Has
the byproduct of stoping the linter fail.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->