Arke-Class Surveying Ship | CLIP-LAND (#3910)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a new specialist ship to CLIP's roster, a small CLIP-LAND ship for
surveying planets and relaying information to friendly ships. Has a
telecommunications setup and basic equipment, as well as a small armory
for the Captain to distribute.


![image](https://github.com/user-attachments/assets/b749ff44-d50d-4059-b6f6-5a574abc2b51)

<details><summary>SDMM Screenshots</summary>
<p>


![image](https://github.com/user-attachments/assets/97a6b133-37b1-48d4-ae95-de031ad5c9d5)


![image](https://github.com/user-attachments/assets/1d589431-3af3-4f41-8747-ffb766c43130)

</p>
</details> 

<details><summary>More In-Game Screenshots</summary>
<p>


![image](https://github.com/user-attachments/assets/ea473637-1ea6-4544-9036-9c15772861a2)

![image](https://github.com/user-attachments/assets/589bbf21-aac9-41be-8bbe-f9b6985b3aa9)

![image](https://github.com/user-attachments/assets/3b3a885c-4b74-4a42-9ba8-2c171f1d4213)

![image](https://github.com/user-attachments/assets/2be483e7-a005-497e-82ab-e0e8dad864cf)

</p>
</details> 
Also adds a tcomms relay to the Atlas, located in the port side of
engineering, as well as giving CLIP headsets to start with.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Adds a new small-crew ship and a new ship for CLIP that isn't part of
CMM. Gives telecommunications access to CLIP ships without the need for
a larger command ship.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Arke-Class Surveying Ship
add: The Atlas-Class now has a Telecomms Relay
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->