Add Hyposprays to Cargo (#4049)
## About The Pull Request
I got to play with one of these and it was cool so I'm adding it to
cargo

Also adds chitosan, a coagulant. since we didn't have a coagulant. for
some reason.
## Why It's Good For The Game

i think its cool :)

## Changelog

:cl:
add: chitosan! a coagulant that you can make by mixing 1 silicon, 2
sugar, 1 cryptobiolin and warming it all up.
add: hyposprays and hypospray vials to cargo
/:cl: