Mech Zoom (#4048)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Changes mech zooming to use the updated version added in #3736. Mech
zooming is now tied to the equipment on the mech.

Reflavours the Solaris to be a beam sniper, and adds zoom to the pulse
cannon.

Removes inherent zoom ability from the Marauder.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

Mechs using the better zooming system is good, and gives the Solaris a
stronger distinct niche.

## Changelog

:cl:
add: Mechs use the new zoom system
add: Solaris and Ez-Pulse cannon can zoom.
del: Zoom on the Marauder
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->