Removes an erroneous airlock from the Derecho (#4063)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

A second locker room airlock was accidentally mapped into the Derecho,
and this PR removes it.

## Why It's Good For The Game

This fixes a mistake in an earlier PR

## Changelog

:cl:
fix: fixed an accidental second locker room airlock in the derecho
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->