Haboob-class Patrol Boat (Small NGR Combat Ship) (#4013)
## About The Pull Request
Adds in the Haboob-class, a new small NGR ship designed primarily for
combat, lacking in other areas. While it does lack in medical and
engineering, it should have the bare minimum of supplies to engage in
some other activities, mostly scrapping or mining. It has a crew
composition of 6, that being the:

Captain
Lieutenant
Mechanic
Operative
Wrecker
Initiate

images:

![image](https://github.com/user-attachments/assets/d46aedd0-0e12-4324-85e5-739e179beb42)

![image](https://github.com/user-attachments/assets/312c42f7-7c55-4e50-bd02-30d89234abc5)
in game:

![image](https://github.com/user-attachments/assets/4a100acc-b488-4fc5-af81-7dc1c5f38004)

![image](https://github.com/user-attachments/assets/333c118a-23e6-4c62-af47-47e18dc791f0)

## Why It's Good For The Game
Small little ship, main goal is to (hopefully) get more people playing
the NGR so it has a reliable playerbase to support its larger ships.
## Changelog

:cl:
add: Added new things

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->