Applies some bugfixes to examine_more interaction with modular computers and adds examine_more interaction to tablets (#3995)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR fixes a few bugs in PR
https://github.com/shiptest-ss13/Shiptest/pull/3553 and it also gives
tablets the ability to be interacted with using the examine_more
feature.

Please let me know how this may perform ingame. I'm still unsure as to
if shuttling this functionality to examine_more is better than applying
it to the examine function.

## Why It's Good For The Game

This PR fixes modular computers from being activated from long range via
hidden bluetooth implants, and it gives tablets the ability to be
interacted with without it having to be in a user's hand, which can be
nice for those that may lay the item down on a table in order to let a
group of people see the screen at once.

## Changelog

:cl:
code: applied some bugfixes to the interaction of examine_more with
modular computers and laptops
add: adds the ability to interact with a tablet using the examine_more
function
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->