Reduces the Komodo to atoms (#4086)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

obliterates, annihilates, and/or eradicates all traces of the Komodo
from the codebase.

## Why It's Good For The Game

We have actual Hardliner ships now, so we can pull the trigger on the
Komodo. Goodbye, don't come back.

## Changelog

:cl:
del: Launched the Komodo-class into a black hole
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->