Fixes AI Movement Runtime (#4100)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Should fix the AI movement runtime that apparently is happening an
absolute ton. AI controllers never removed themselves from the movement
processing lists when deleted, causing them to get hard deleted and
replaced with nulls in the list. This fixes that.

Supersedes #4082

## Why It's Good For The Game
wowie that's a lotta runtimes

## Changelog

:cl:
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->