Minor Abandoned Listening Post Housekeeping (#4096)
## About The Pull Request

Adjusts the abandoned listening post minorly after it's introduction. 

Fixes firearms not being in their cabinets, replaces the Cobra with a
Sidewinder, adds a Cybersun uniform in a closet, and a stamp on a table.

## Why It's Good For The Game

Brings this ruin a bit more in-line with others, and doesn't make it's
loot feel so nothingburger, especially after the Mob Armor PR.

## Changelog

:cl:
balance: Abandoned Listening Post has been slightly adjusted to have
marginally better loot.
/:cl: