The floral shirt now takes two seconds to put on. (#4129)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

read title

## Why It's Good For The Game
because a winter coat taking less time than a shirt that's barely even
being worn is bullshit.
yes. this is a spite pr.
## Changelog

:cl:
fix: Floral shirts no longer take six whole seconds to put on, because
that is completely insane.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->