Added Hardliner phorid envirosuits. (#4177)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Added sprites for a Hardliner Phorid Envirosuit and Helmet, and added
them together as an 'outfit.'

NOTE: This doesn't change any round-start gear for Phorids as that's its
own PR entirely. The only way to acquire this would be to have an admin
spawn it in for you, similar to the NGR Phorid Envirosuit + Helmet.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The more faction themed gear, the better! (Even if we can't access it
without admin help.)
With more Phorid gear, the likely-hood of someone getting around to
changing how their outfits are distributed inches closer.

![icon0](https://github.com/user-attachments/assets/37f754b2-f82e-4388-aaa6-bb72b725bf31)

![icon1](https://github.com/user-attachments/assets/d2fe23e8-251a-4cfc-9edc-dd001b52509e)

![icon2](https://github.com/user-attachments/assets/a85313d7-373e-4bd5-918f-bf42b98255f3)

![pose0](https://github.com/user-attachments/assets/56d3a76c-70aa-4e57-b95b-7e6367863843)

![pose1](https://github.com/user-attachments/assets/0ca03260-3b8c-4662-823b-b8c26fdbf022)

![pose2](https://github.com/user-attachments/assets/c7c9ff2f-65c2-4e3e-8a37-b1d2214f9fa9)

![pose3](https://github.com/user-attachments/assets/61905265-4812-435c-b87c-5bdd5d643690)

![pose4](https://github.com/user-attachments/assets/16dc5a1e-f0bf-4148-9232-29f4b3ffcb5a)

![pose5](https://github.com/user-attachments/assets/51439ab3-e658-43e1-b6fe-4da0dbbd07d9)

![pose6](https://github.com/user-attachments/assets/72c1a483-8859-4e70-9ff0-599b39a31ec5)

![pose7](https://github.com/user-attachments/assets/d5ca080f-96bd-44b6-816c-1052ad021c35)

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added Hardliner phorid envirosuit and helmet sprites.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->