Hand Labeler is now Small + Paper Rolls and Evidence Bags in the Autolathe (#4210)
## About The Pull Request

Hand Labeler is now small, so it can fit in bags and other storage
spaces (like cabinets). Paper rolls are now in the autolathe. So are
evidence bags.

## Why It's Good For The Game

Hand Labelers (bureaucracy tool for labeling things) really shouldn't
be... the size of boxes and weapons and the like. They're not super
powerful and honestly could fit in a bag or a cabinet quite easily.
Paper refill rolls means you don't have to dump 50,000 of these things
as waste products.

Evidence bags are on a side note something that should be available for
storing small items securely and sorting them. I've heard it floated for
organs and similar.

## Changelog

:cl:
balance: Hand Labelers are now small.
add: Paper Rolls and Evidence Bags are now in the autolathe.
/:cl:

---------

Signed-off-by: firebudgy <153147550+firebudgy@users.noreply.github.com>