Parity for the Colossus, Talos, Valor and Vaquero (#4175)
## About The Pull Request

Slightly buffs the armories of the Colossus and Vaquero to bring them up
to par with most modern combat-based ships.

Adds a singular Mastiff, a bit of ammo and 2 magazines for it onto the
Colossus's armory. Removes 3,000 from roundstart funding in exchange and
replaces a Commissioner.

Also adds a Rattlesnake + 2 magazines to the Vaquero. This replaces a
Commissioner.

Third additions are Pinschers (the designated officer pistol for Inteq,
as intended in the original M20 Auto Elite PR) to the Vanguard quarters
of all Vanguards.

## Why It's Good For The Game

This brings these ships up to par with others within their class, and
means that the designated go outside roles have enough long arms.

## Changelog

:cl:
balance: Rattlesnake added to the Vaquero and Talos, Mastiff added to
the Colossus.
add: Pinschers to all Vanguards.
/:cl: