Minor Kali Andhi and Derecho Equipment Tweaks (#4200)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds a surgical case and anesthetics to the Kali Andhi, and it
adds half masks to the lockers for the Wreckers and the Mechanic aboard
the Derecho.

## Why It's Good For The Game

Surgical equipment is nice to have on the Kali Andhi, and it allows the
onboard doctor to be able to do more roundstart. Halfmasks are a nice
and stylish secondary option for NGR characters to use for their
internals.

## Changelog

:cl:
add: Added surgical equipment to the Kali Andhi
add: Added additional half masks for the wreckers and the mechanic
aboard the Derecho
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->