Adds Proper Digitigrade Sprites To Syndicate Hardsuits (#4208)
## About The Pull Request

Adds digitigrade sprites to some syndicate suits.
Base, beige-red, pointman, rusted, and white-red.

<details>
<summary>Pictures</summary>


![image](https://github.com/user-attachments/assets/395a1795-f01f-4c0a-afe8-157d248fe908)


![realdigihardsuit](https://github.com/user-attachments/assets/ef6311a5-a041-40c2-a3c9-d82287d27aef)

</details>

## Why It's Good For The Game

Allows for srothy to look normal in hardsuits, because I was SICK OF
IT!!! Be the change you want in the world folks.
May change this to all unsupported hardsuits, but it's really not much
outside of these.

## Changelog

:cl:
add: New digitigrade hardsuit sprites
/:cl: