Atmospheres Refactor Take 2 (#4099)
## About The Pull Request

Re-attempt of #3878.

Basically makes atmosphere gas string generation a lot better. The
previous method was baffling to me and it bugged me to no end. This was
intended to be part of a bigger PR but I might as well PR this as it is
now.

Calculations of randomized atmos mixes **WILL** be different now, but I
tried to make it as similar to the old method as I could without bending
over backwards. We'll get used to it, unless it's bugged.

Also gets rid of `initial_temperature` on turfs, because it literally
did nothing. Specify temperature as part of `initial_gas_mix`.

## Why It's Good For The Game

Simplifies an EXTREMELY needlessly complicated process that did
something as simple as generating a "random" gas string for planets to
use. Allows me to feel less bad when I generate a lot more unique
atmosphere types.

## Changelog

:cl:
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->