.50BMG Ammo Boxes (plus a small rebalance to .50) (#4161)
## About The Pull Request

Adds one of the last ammo boxes we were missing, for .50BMG - sniper
rounds for the Taipan. They are in cargo for 1000 credits for 20. Given
you're operating an AMR and it's already ruin loot, this should keep
it's kills you factor balanced.

It no longer delimbs and also has halfed stun. But given that it's as
powerful as it is, it shouldn't impact it too much outside of it being
an "I win".

The Taipan itself is not added to cargo, nor are it's magazines. Simply
the ammo.

Imaginos also wanted this for some NGR RP and I agreed.

Ammo box sprites done by @rye-rice! Thanks Rylie!

## Why It's Good For The Game

People might actually use the Taipan instead of treating it as a trophy
from the single ruin it's in, and also makes it slightly less a
turbo-win button since that's not interesting in the slightest.

## Changelog

:cl: Cloudbreak, Ryerice
add: 50BMG ammo boxes to cargo for 1000 credits.
balance: 50BMG doesn't instantly delimb you anymore.
/:cl:

---------

Signed-off-by: firebudgy <153147550+firebudgy@users.noreply.github.com>
Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>