Reduces spawn limit for Mudskipper and Kilo (#4248)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Mudskipper and Kilo were the only ships with a spawn limit higher than
1, and this was to help ensure smaller ships would be available for
players to spawn. Now that several small independent ships are being
added to the game, this is less of a concern. Also it might reduce the
number of solo ships clogging up the manifest, marginally.

## Why It's Good For The Game

maybe a little less soloshipping would be good for the game. you WILL
interact with other players in the multiplayer game

## Changelog

:cl:
balance: Reduced Mudskipper and Kilo spawn limit to 1 each
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->