Dynamic Gen VI Mobs (replaces crusty NT guys (#4190)
## About The Pull Request

to-do, fill in

## Why It's Good For The Game


![dreamseeker_RlTFYLIffo](https://github.com/user-attachments/assets/95b4d23e-aaba-4d88-9544-a9182359e543)


## Changelog

:cl:
add: replaces old Nanotrasen Security with dynamic gen mobs
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->